refactored placeholder entity 'Design2' to 'Design', commented out old Design
