--- src/main/java/nl/novi/be_plantjesplanner/dtos/Design2Dto.java
@@ -1,4 +1,0 @@
-package nl.novi.be_plantjesplanner.dtos;
-
-public record Design2Dto(Long id, String title) {
-}
