+++ src/main/java/nl/novi/be_plantjesplanner/dtos/DesignDto.java
@@ -1,0 +1,4 @@
+package nl.novi.be_plantjesplanner.dtos;
+
+public record DesignDto(Long id, String title) {
+}
