+++ src/main/java/nl/novi/be_plantjesplanner/dtos/SelectedPlantDto.java
@@ -1,6 +1,3 @@
 package nl.novi.be_plantjesplanner.dtos;
 
-import nl.novi.be_plantjesplanner.entities.Design;
-import nl.novi.be_plantjesplanner.entities.Plant;
-
-public record SelectedPlantDto(Long id, Integer quantity, Design2Dto design2Dto, PlantDto plantDto){}
+public record SelectedPlantDto(Long id, Integer quantity, DesignDto designDto, PlantDto plantDto){}
