--- src/main/java/nl/novi/be_plantjesplanner/entities/Design2.java
@@ -1,35 +1,0 @@
-package nl.novi.be_plantjesplanner.entities;
-
-import jakarta.persistence.*;
-@Entity
-@Table(name = "design2")
-public class Design2 {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String title;
-
-    @OneToOne(mappedBy = "design2")
-    private User user;
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public User getUser() {
-        return user;
-    }
-
-    public void setUser(User user) {
-        this.user = user;
-    }
-}
